refactoring submit_aes_request method
